+++ src/web-api.js
@@ -6,23 +6,7 @@ const JSzip = require('jszip');
 const Signature = require('./signature');
 const Utilities = require('./utilities');
 const IDApi = require('./id-api');
-const { mapServerUri, sdkVersionInfo } = require('./helpers');
-
-const validatePartnerParams = (partnerParams) => {
-  if (!partnerParams) {
-    throw new Error('Please ensure that you send through partner params');
-  }
-
-  if (typeof partnerParams !== 'object') {
-    throw new Error('Partner params needs to be an object');
-  }
-
-  ['user_id', 'job_id', 'job_type'].forEach((key) => {
-    if (!partnerParams[key]) {
-      throw new Error(`Please make sure that ${key} is included in the partner params`);
-    }
-  });
-};
+const { mapServerUri, sdkVersionInfo, validatePartnerParams } = require('./helpers');
 
 class WebApi {
   constructor(partner_id, default_callback, api_key, sid_server) {
